+++ src/main/kotlin/app/revanced/patches/reddit/customclients/syncforreddit/api/fingerprints/GetAuthorizationStringFingerprint.kt
@@ -1,0 +1,7 @@
+package app.revanced.patches.reddit.customclients.syncforreddit.api.fingerprints
+
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+
+object GetAuthorizationStringFingerprint : MethodFingerprint(
+    strings = listOf("authorize.compact?client_id")
+)