+++ src/main/kotlin/app/revanced/patches/syncforreddit/api/fingerprints/GetAuthorizationStringFingerprint.kt
@@ -1,7 +1,0 @@
-package app.revanced.patches.syncforreddit.api.fingerprints
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object GetAuthorizationStringFingerprint : MethodFingerprint(
-    strings = listOf("authorize.compact?client_id")
-)