+++ src/main/kotlin/app/revanced/patches/syncforreddit/api/fingerprints/GetBearerTokenFingerprint.kt
@@ -1,7 +1,0 @@
-package app.revanced.patches.syncforreddit.api.fingerprints
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object GetBearerTokenFingerprint : MethodFingerprint(
-    strings = listOf("Basic")
-)