+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/comments/patch/CommentsPatch.kt
@@ -9,21 +9,23 @@ import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
-import app.revanced.patches.shared.mapping.misc.patch.ResourceMappingPatch
 import app.revanced.patches.shared.settings.preference.impl.PreferenceScreen
 import app.revanced.patches.shared.settings.preference.impl.StringResource
 import app.revanced.patches.shared.settings.preference.impl.SwitchPreference
 import app.revanced.patches.youtube.layout.hide.comments.annotations.HideCommentsCompatibility
+import app.revanced.patches.youtube.misc.litho.filter.patch.LithoFilterPatch
 import app.revanced.patches.youtube.misc.settings.bytecode.patch.SettingsPatch
 
 @Patch
 @Name("comments")
 @Description("Hides components related to comments.")
 @HideCommentsCompatibility
-@DependsOn([SettingsPatch::class, ResourceMappingPatch::class])
+@DependsOn([SettingsPatch::class, LithoFilterPatch::class])
 @Version("0.0.1")
 class CommentsPatch : ResourcePatch {
     override fun execute(context: ResourceContext): PatchResult {
+        LithoFilterPatch.addFilter(FILTER_CLASS_DESCRIPTOR)
+        
         SettingsPatch.PreferenceScreen.LAYOUT.addPreferences(
             PreferenceScreen(
                 "revanced_comments_preference_screen",
@@ -47,5 +49,9 @@ class CommentsPatch : ResourcePatch {
         )
 
         return PatchResultSuccess()
+    }
+    
+    private companion object {
+        const val FILTER_CLASS_DESCRIPTOR = "Lapp/revanced/integrations/patches/components/CommentsFilter;"
     }
 }